third_party/blink/web_tests/external/wpt/service-workers/cache-storage/cache-match.https.any.serviceworker-expected.txt

This is a testharness.js-based test.
[FAIL] MIME type should reflect Content-Type headers of response.
  assert_equals: mime type can be overridden expected "text/plain" but got "text/html"
Harness: the test ran to completion.
